Renamed userResolvers to userQueries, as the file will be specific to user related queries, while a new file called userMutations can be created for the Mutation object.
