+++ resolvers/index.js
@@ -1,10 +1,10 @@
-const userResolvers = require("./userResolvers");
+const userQueries = require("./userQueries");
 const typeResolvers = require("./typeResolvers");
 
 const resolvers = {
 	Query: {
 		info: () => `This is the API of CleanLift`,
-		...userResolvers
+		...userQueries
 	},
 	// Mutation: {}, Uncomment and add mutation resolvers inside.
 	...typeResolvers
